+++ src/hooks/accounts/useAccount.tsx
@@ -1,8 +1,8 @@
 import useSWR from 'swr'
 
 import getAccount from 'api/accounts/getAccount'
-import getV1Positions from 'api/v1/getV1Positions'
 import useChainConfig from 'hooks/useChainConfig'
+import useV1Account from 'hooks/v1/useV1Account'
 import useStore from 'store'
 
 export default function useAccount(accountId?: string, suspense?: boolean) {
@@ -10,16 +10,17 @@ export default function useAccount(accountId?: string, suspense?: boolean) {
   const address = useStore((s) => s.address)
   const isV1 = accountId === address
 
-  const cacheKey = isV1
-    ? `chains/${chainConfig.id}/v1/user/${accountId}`
-    : `chains/${chainConfig.id}/accounts/${accountId}`
-
-  return useSWR(
-    accountId && cacheKey,
-    () => (isV1 ? getV1Positions(chainConfig, accountId) : getAccount(chainConfig, accountId)),
+  const v1Account = useV1Account(accountId, suspense)
+  const v2Account = useSWR(
+    !!accountId && !isV1 && `chains/${chainConfig.id}/accounts/${accountId}`,
+    () => getAccount(chainConfig, accountId),
     {
       suspense: suspense,
       revalidateOnFocus: false,
     },
   )
+
+  if (isV1) return v1Account
+
+  return v2Account
 }
