+++ src/hooks/v1/useV1Deposits.ts
@@ -5,11 +5,13 @@ import useChainConfig from 'hooks/useChainConfig'
 import useStore from 'store'
 
 export default function useV1Deposits() {
+  const isV1 = useStore((s) => s.isV1)
+
   const chainConfig = useChainConfig()
   const address = useStore((s) => s.address)
 
   return useSWR(
-    !!address && `${chainConfig.id}/v1/deposits/${address}`,
+    !!address && isV1 && `${chainConfig.id}/v1/deposits/${address}`,
     () => getV1Deposits(chainConfig, address ?? ''),
     {
       suspense: true,
